requirements.txt
# Django and REST Framework
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9

# Authentication and Security
python-dotenv>=1.0.0
django-environ>=0.11.2

# Email
django-sendgrid-v5>=0.8.1
celery>=5.3.0
redis>=5.0.0

# Development Tools
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.1
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0

# Documentation
drf-yasg>=1.21.7 